[IMPLEMENTATION][REFACTORING] -> minor changes to classes, trying to implement INSERT queries for customers

+++ management_system/src/main/sql/create-table-scheme.sql
@@ -33,7 +33,7 @@ CREATE TABLE Day
 (
     day_id_                     SERIAL              PRIMARY KEY,
     month_id_                   INTEGER             NOT NULL REFERENCES  Month(month_id_),
-    time_worked_                INTEGER             NOT NULL,
+    time_worked_                NUMERIC(4, 2)       NOT NULL,
     date_                       DATE                NOT NULL,
     type_                       VARCHAR(32)         NOT NULL
 );
